Move `ParametersKSerializer` to `headers` package
